Post-Jump slowdown replaces windup (with skill checks) (#146)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Puts jump windup on a config option, replacing it with post-jump
slowdown.

I've addressed balance concerns leveled in [this conversation on the
discord.](https://ptb.discord.com/channels/1393132000583618701/1396599047128617010/1423933973729509379)

When two characters collide, both the jumper and the victim make a
contested roll. If the victim succeeds, they do not get knocked down or
disarmed.

The winner is calculated thusly:
We check the jumper's Strength + Brawl and Dex + Athletics and keep the
higher of the two values. We then check the victim's Stamina + Athletics
or Stamina + Brawl, whichever's higher. Finally, we roll these values
against a median difficulty of 6 (as per medium difficulty in V20 and
the standard V5 rolling rules). The owner of the higher value wins. In
cases of a tie, the jumper wins. The loser is paralyzed, knocking them
down and disarming them.

Regardless of whether or not the jumper impacts something during the
jump, they are slowed in such a way that it would have been approx. 1
tile faster to have walked that distance assuming they wish to continue
moving in that direction. This both punishes jumping as a casual
traversal method and provides a window for agile opponents to punish the
jumper if they are jumping recklessly.

The latter change was implemented on Apocrypha XIII in
https://github.com/ApocryphaXIII/ApocryphaXIII/pull/183 and has had
extensive testing, of which feedback was overwhelmingly positive. This
change will also be tied to a config option.

Thank you to @hex37 for discussing these balance concerns with me and
originally implementing jumping to SecondCity in #75 and #98.

Depends on #144 



https://github.com/user-attachments/assets/91fc443f-d5ff-4d95-8a09-503c14486bb5


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Players on Apocrypha XIII and a few members of The Final Nights have
come to me variously regarding the [aforementioned
PR](https://github.com/ApocryphaXIII/ApocryphaXIII/pull/183) with
extremely strong opinions AGAINST jump windup. That being said, Requiem
still seems to desire jump windup so it has been left as a config option
if it is desired.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Replaced jump windup with post-jump slowdown
code: Added skillchecks to carbon throw_impact code
config: Jump windup and jump slowdown config options
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>